Move <nav> and common <head> elements to includes
The default template is rather restrictive. Jekyll doesn't allow
creating templates blocks the way Jinja does; if we want to add
anything in <head> we have to go through the layout file.

Case in point: crossfire needs a different meta viewport.
Now that viewport logic no longer needs to be part of a page's
front matter, it's contained in the head fragment.